fix extension

+++ Korenko HM 24/Korenko HM 24/Extension/Int+Extension.swift
@@ -9,7 +9,7 @@ import Foundation
 import UIKit
 import RealmSwift
 
-extension Double{
+extension Int{
   
     var timeHHmm: String {
         let date = Date(timeIntervalSince1970: self)
